backend/run.py
"""
Application entry point.
Run with: flask run  (respects FLASK_ENV and FLASK_APP env vars)
Or directly: python run.py
"""
import os
from dotenv import load_dotenv

load_dotenv()

from app import create_app

app = create_app(os.environ.get('FLASK_ENV', 'development'))

if __name__ == '__main__':
    app.run(
        host='0.0.0.0',
        port=int(os.environ.get('PORT', 5000)),
        debug=app.config.get('DEBUG', False),
    )
