backend/requirements.txt
Flask>=3.0
Flask-Bcrypt>=1.0
Flask-WTF>=1.2
Flask-Session>=0.6
Flask-SQLAlchemy>=3.1
Flask-Limiter>=3.5
Flask-CORS>=4.0
PyOTP>=2.9
qrcode[pil]>=7.4
twilio>=9.0
python-dotenv>=1.0
itsdangerous>=2.1
cryptography>=41.0
requests>=2.31
pytest>=8.0
pytest-flask>=1.3
